Modularized the code
