+++ frontend/src/components/HowItWorksSection.jsx
@@ -1,131 +1,13 @@
 import React, { useState, useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { PlusSquare, Search, UserPlus, CheckCircle, ArrowRight } from "lucide-react";
+import { ArrowRight } from "lucide-react";
+import { steps } from './stepsData';
+import StepCard from './StepCard';
 
-const steps = [
-  {
-    icon: UserPlus,
-    title: "Create an Account",
-    description: "Sign up for free to get your personalized profile and start your collection.",
-    bgColor: "from-red-500 to-pink-600",
-    features: ["Free forever", "Secure profile", "Personal dashboard"]
-  },
-  {
-    icon: PlusSquare,
-    title: "Add Your Recipes",
-    description: "Share your favorite dishes with our community using our simple recipe editor.",
-    bgColor: "from-yellow-500 to-orange-600",
-    features: ["Easy editor", "Photo uploads", "Step-by-step guide"]
-  },
-  {
-    icon: Search,
-    title: "Explore & Interact",
-    description: "Discover new meals, leave comments, and like the recipes that inspire you.",
-    bgColor: "from-green-500 to-emerald-600",
-    features: ["Smart search", "Save favorites", "Community interaction"]
-  }
-];
-
-const StepCard = ({ step, index, isActive, onHover, onIconClick }) => {
-  const IconComponent = step.icon;
-
-  return (
-    <div
-      className={`relative p-8 rounded-2xl transition-all duration-700 transform hover:-translate-y-4 ${
-        isActive
-          ? 'bg-white dark:bg-slate-800 shadow-2xl scale-105'
-          : 'bg-white/50 dark:bg-slate-800/50 shadow-lg hover:shadow-xl'
-      }`}
-      onMouseEnter={() => onHover(index)}
-      onMouseLeave={() => onHover(-1)}
-      style={{ animationDelay: `${index * 200}ms` }}
-    >
-      {/* Icon (merged version) */}
-      <div className="mb-6 flex justify-center">
-        {onIconClick ? (
-          <button
-            onClick={onIconClick}
-            className={`relative w-20 h-20 rounded-2xl bg-gradient-to-br ${step.bgColor} shadow-lg flex items-center justify-center transform transition-all duration-500 ${
-              isActive ? 'scale-110 rotate-3' : ''
-            }`}
-            aria-label={step.title}
-            type="button"
-          >
-            <IconComponent className="w-10 h-10 text-white" />
-            <div
-              className={`absolute inset-0 bg-gradient-to-br ${step.bgColor} rounded-2xl blur-xl opacity-50 ${
-                isActive ? 'opacity-75' : 'opacity-0'
-              } transition-opacity duration-500`}
-            />
-          </button>
-        ) : (
-          <div
-            className={`relative w-20 h-20 rounded-2xl bg-gradient-to-br ${step.bgColor} shadow-lg flex items-center justify-center transform transition-all duration-500 ${
-              isActive ? 'scale-110 rotate-3' : ''
-            }`}
-          >
-            <IconComponent className="w-10 h-10 text-white" />
-            <div
-              className={`absolute inset-0 bg-gradient-to-br ${step.bgColor} rounded-2xl blur-xl opacity-50 ${
-                isActive ? 'opacity-75' : 'opacity-0'
-              } transition-opacity duration-500`}
-            />
-          </div>
-        )}
-      </div>
-
-      {/* Step Number */}
-      <div className="absolute -top-4 -left-4 w-12 h-12 bg-gradient-to-r from-gray-800 to-gray-700 dark:from-white dark:to-gray-200 rounded-full flex items-center justify-center border-4 border-white dark:border-slate-900 shadow-lg">
-        <span className="text-white dark:text-gray-800 font-bold text-lg">{index + 1}</span>
-      </div>
-
-      {/* Content */}
-      <div className="relative z-10 text-center">
-        <h3 className="text-xl font-bold mb-4 text-gray-800 dark:text-white">
-          {step.title}
-        </h3>
-        <p className="text-gray-600 dark:text-gray-300 mb-6 leading-relaxed">
-          {step.description}
-        </p>
-
-        <div className="space-y-2">
-          {step.features.map((feature, featureIndex) => (
-            <div key={featureIndex} className="flex items-center justify-center gap-2 text-sm text-gray-500 dark:text-gray-400">
-              <CheckCircle className="w-4 h-4 text-green-500" />
-              <span>{feature}</span>
-            </div>
-          ))}
-        </div>
-      </div>
-
-      {/* Hover Arrow */}
-      <button
-  type="button"
-  onClick={(e) => {
-    e.stopPropagation();
-    if (onIconClick) onIconClick();
-  }}
-  onKeyDown={(e) => {
-    if ((e.key === "Enter" || e.key === " ") && onIconClick) {
-      e.preventDefault();
-      onIconClick();
-    }
-  }}
-  aria-label={onIconClick ? `Go to ${step.title}` : undefined}
-  className={`absolute bottom-6 right-6 w-10 h-10 bg-gradient-to-r ${step.bgColor} rounded-full flex items-center justify-center transition-all duration-300 ${
-    isActive ? 'opacity-100 scale-100' : 'opacity-0 scale-50'
-  } ${onIconClick ? 'cursor-pointer' : 'cursor-default'} z-20`}
->
-  <ArrowRight className="w-5 h-5 text-white" />
-</button>
-    </div>
-  );
-};
 const HowItWorksSection = () => {
   const navigate = useNavigate();
   const [activeStep, setActiveStep] = useState(-1);
   const [currentStep, setCurrentStep] = useState(0);
-
 
   useEffect(() => {
     const interval = setInterval(() => {
@@ -136,7 +18,7 @@ const HowItWorksSection = () => {
 
   return (
     <section id="about" className="py-20 bg-gradient-to-br from-slate-50 via-white to-slate-100 dark:from-slate-900 dark:via-slate-800 dark:to-slate-900 relative overflow-hidden">
-      {/* Animated Background Elements */}
+      {/* Animated Background */}
       <div className="absolute inset-0 overflow-hidden pointer-events-none">
         <div className="absolute top-32 left-16 w-64 h-64 bg-gradient-to-r from-red-200 to-pink-200 rounded-full blur-3xl opacity-20 animate-pulse" />
         <div className="absolute bottom-32 right-16 w-80 h-80 bg-gradient-to-r from-yellow-200 to-orange-200 rounded-full blur-3xl opacity-15 animate-bounce" style={{ animationDelay: '1s' }} />
@@ -150,7 +32,6 @@ const HowItWorksSection = () => {
             <div className="w-3 h-3 rounded-full bg-gradient-to-r from-blue-500 to-cyan-600 animate-pulse mr-3" />
             <span className="text-sm font-semibold text-blue-600 dark:text-blue-400">How It Works</span>
           </div>
-          
           <h2 className="text-4xl md:text-5xl lg:text-6xl font-black text-gray-800 dark:text-white mb-6">
             Get Cooking in{' '}
             <span className="relative">
@@ -160,7 +41,6 @@ const HowItWorksSection = () => {
               <div className="absolute -bottom-2 left-0 right-0 h-1 bg-gradient-to-r from-red-500 via-yellow-500 to-green-500 rounded-full transform scale-x-0 group-hover:scale-x-100 transition-transform duration-500" />
             </span>
           </h2>
-          
           <p className="text-xl text-gray-600 dark:text-gray-300 max-w-4xl mx-auto leading-relaxed">
             Join our vibrant community and start your culinary journey today. It's simple, fun, and completely free!
           </p>
@@ -169,31 +49,21 @@ const HowItWorksSection = () => {
         {/* Steps Grid */}
         <div className="grid grid-cols-1 md:grid-cols-3 gap-8 mb-16">
           {steps.map((step, index) => (
-  <StepCard 
-    key={index} 
-    step={step} 
-    index={index}
-    isActive={activeStep === index || currentStep === index}
-    onHover={setActiveStep}
-    onIconClick={
-      step.title === "Create an Account"
-        ? () => navigate('/register')
-        : step.title === "Add Your Recipes"
-        ? () => navigate('/recipes')
-        : undefined
-    }
-  />
-))}
-
-        </div>
-
-        {/* Connecting Lines (Desktop) */}
-        <div className="hidden md:block absolute top-1/2 left-1/2 transform -translate-x-1/2 -translate-y-1/2 w-full max-w-4xl pointer-events-none">
-          <div className="relative">
-            <div className="absolute top-0 left-1/6 right-1/6 h-0.5 bg-gradient-to-r from-red-300 via-yellow-300 to-green-300 opacity-30" />
-            <div className="absolute top-0 left-1/6 w-8 h-0.5 bg-gradient-to-r from-red-500 to-yellow-500 animate-pulse" />
-            <div className="absolute top-0 right-1/6 w-8 h-0.5 bg-gradient-to-r from-yellow-500 to-green-500 animate-pulse" style={{ animationDelay: '1s' }} />
-          </div>
+            <StepCard 
+              key={index} 
+              step={step} 
+              index={index}
+              isActive={activeStep === index || currentStep === index}
+              onHover={setActiveStep}
+              onIconClick={
+                step.title === "Create an Account"
+                  ? () => navigate('/register')
+                  : step.title === "Add Your Recipes"
+                  ? () => navigate('/recipes')
+                  : undefined
+              }
+            />
+          ))}
         </div>
 
         {/* Bottom CTA */}
